schemas/requirements.txt
avro-python3
confluent_kafka
requests
wget
